попытка всё починить номер 2
